+++ src/main/java/de/maxhenkel/camera/gui/CameraScreen.java
@@ -11,8 +11,6 @@ import net.minecraft.util.text.TranslationTextComponent;
 
 import java.awt.image.BufferedImage;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.UUID;
 
 public class CameraScreen extends ContainerScreen {
@@ -80,8 +78,13 @@ public class CameraScreen extends ContainerScreen {
                     minecraft.currentScreen = null;
                 });
             }));
-            upload.active = ImageTools.isFileChooserAvailable();
         }
+    }
+
+    @Override
+    public void tick() {
+        super.tick();
+        upload.active = !ImageTools.isFileChooserOpen();
     }
 
     private void sendShader() {
@@ -103,12 +106,6 @@ public class CameraScreen extends ContainerScreen {
         int shaderWidth = font.getStringWidth(shaderName);
 
         font.drawStringWithShadow(shaderName, xSize / 2 - shaderWidth / 2, ySize / 2 - font.FONT_HEIGHT / 2, 0xFFFFFFFF);
-
-        if (upload != null && upload.isHovered() && !ImageTools.isFileChooserAvailable()) {
-            List<String> list = new ArrayList<>();
-            list.add(new TranslationTextComponent("message.camera.no_java_fx").getUnformattedComponentText());
-            renderTooltip(list, x - guiLeft, y - guiTop);
-        }
     }
 
     @Override